fix: use unsafe.Pointer for HashMap base, enforce off-heap value contract
mapState.base changed from uintptr to unsafe.Pointer with single-expression
pointer arithmetic to satisfy checkptr instrumentation under -race. All
HashMap tests and benchmarks now use Arena-allocated off-heap values instead
of Go heap pointers, consistent with every other allocator in the package.

The GC never scans the mmap'd bucket array — Go heap pointers stored in the
map are invisible to the collector. Doc comments and README now document
this contract explicitly.

+++ hashmap_helpers.go
@@ -9,7 +9,7 @@ import "unsafe"
 func (h *HashMap) offHeapPtr(bucketIdx uint64) unsafe.Pointer {
 	// bucketIdx is constrained by (size - 1) prior to calling.
 	s := h.state.Load()
-	return unsafe.Pointer(s.base + uintptr(bucketIdx)*128)
+	return unsafe.Pointer(uintptr(s.base) + uintptr(bucketIdx*128))
 }
 
 // nextPow2 mathematically rounds up a 64-bit integer to the nearest power of 2.
